refactor: 패키지 통일화
