+++ history-service/src/main/java/com/study/history/service/entity/History.java
@@ -1,4 +1,4 @@
-package com.study.history.service.domain;
+package com.study.history.service.entity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
