+++ history-service/src/main/java/com/study/history/service/response/StockResponse.java
@@ -1,6 +1,6 @@
 package com.study.history.service.response;
 
-import com.study.history.service.domain.History;
+import com.study.history.service.entity.History;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
