+++ history-service/src/main/resources/application.yml
@@ -1,3 +1,6 @@
+server:
+  port: 8083
+
 spring:
   datasource:
     url: jdbc:mysql://localhost:3306/study_history?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC&useLegacyDatetimeCode=false
@@ -21,3 +24,7 @@ spring:
     init:
       mode: always
       data-locations: classpath:data-init.sql
+#
+#  h2:
+#    console:
+#      enabled: false
